@@ -2,7 +2,7 @@
     "name": "shardj/zf1-future",
     "description": "Zend Framework 1. The aim is to keep ZF1 working with the latest PHP versions",
     "type": "library",
-    "version": "1.24.2",
+    "version": "1.25.0",
     "keywords": [
         "framework",
         "zf1"
@@ -10,16 +10,19 @@
     "homepage": "https://github.com/Shardj/zf1-future/",
     "license": "BSD-3-Clause",
     "require": {
-        "php": ">=7.1",
-        "symfony/polyfill-php81": "^1.24",
-        "symfony/polyfill-ctype": "^1.27",
-        "symfony/polyfill-mbstring": "^1.26"
+        "php": ">=7.4",
+        "symfony/polyfill-ctype": "^1.30",
+        "symfony/polyfill-mbstring": "^1.30",
+        "symfony/polyfill-php81": "^1.30",
+        "symfony/polyfill-php82": "^1.30",
+        "symfony/polyfill-php83": "^1.30"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3",
         "php-parallel-lint/php-parallel-lint": "^1.3",
+        "phpstan/phpstan": "^1.4 || ^2.0",
         "phpunit/phpunit": "^7|^8|^9",
-        "rector/rector": "0.12.19 || ^1.2",
+        "rector/rector": "0.12.19 || ^1.2 || ^2.0",
         "yoast/phpunit-polyfills": "^2.0"
     },
 
@@ -53,8 +56,10 @@
         "lint": "parallel-lint --exclude vendor --exclude tests/Zend/Loader/_files/ParseError.php .",
         "php-cs-fixer:test": "bin/php-cs-fixer fix --dry-run --diff",
         "php-cs-fixer:fix": "bin/php-cs-fixer fix",
-        "rector:test": "bin/rector process --dry-run",
-        "rector:fix": "bin/rector process",
+        "phpstan": "XDEBUG_MODE=off php bin/phpstan analyze --configuration .phpstan.dist.neon",
+        "phpstan:baseline": "XDEBUG_MODE=off php bin/phpstan analyze --configuration .phpstan.dist.neon -b .phpstan.dist.baseline.neon",
+        "rector:test": "bin/rector process --config .rector.php --dry-run",
+        "rector:fix": "bin/rector process --config .rector.php",
         "test": [
             "Composer\\Config::disableProcessTimeout",
             "phpunit --bootstrap tests/TestHelper.php tests/Zend/AllTests.php"
@@ -63,6 +68,8 @@
     "scripts-descriptions": {
         "php-cs-fixer:test": "Run php-cs-fixer",
         "php-cs-fixer:fix": "Run php-cs-fixer and fix issues",
+        "phpstan": "Run phpstan",
+        "phpstan:baseline": "Run phpstan and update baseline",
         "rector:test": "Run rector",
         "rector:fix": "Run rector and fix issues"
     },
