@@ -1,10 +1,12 @@
 <?php
 
+use LimeSurvey\Models\Services\CopySurveyOptions;
 use LimeSurvey\Models\Services\CopySurveyResources;
 use LimeSurvey\Models\Services\FileUploadService;
 use LimeSurvey\Models\Services\FilterImportedResources;
 use LimeSurvey\Models\Services\GroupHelper;
 use LimeSurvey\Models\Services\SurveyAccessModeService;
+use LimeSurvey\Models\Services\SurveyDetailService;
 
 /**
  * Class SurveyAdministrationController
@@ -66,6 +68,7 @@ class SurveyAdministrationController extends LSBaseController
             'postOnly + copy'
         ];
     }
+
     /**
      * SurveyAdministrationController constructor.
      * @param $id
@@ -422,7 +425,7 @@ class SurveyAdministrationController extends LSBaseController
         $arrayed_data['title_bar']['title'] = gT('New survey');
 
         // topbar
-        $aData['topbar']['title'] = gT('Create or copy survey');
+        $aData['topbar']['title'] = gT('Create survey');
         $aData['topbar']['rightButtons'] = $this->renderPartial(
             'partial/topbarBtns_create_survey/rightSideButtons',
             [],
@@ -2044,7 +2047,6 @@ class SurveyAdministrationController extends LSBaseController
             $aData['moreInfo'] = $temp;
         }
 
-        App()->getClientScript()->registerScriptFile(App()->getConfig('adminscripts') . 'surveysettings.js');
         App()->getClientScript()->registerPackage('jquery-json');
 
         // override survey settings if global settings exist
@@ -2207,11 +2209,12 @@ class SurveyAdministrationController extends LSBaseController
     }
 
     /**
-     * Function responsible to import/copy a survey based on $action.
+     * Function responsible to copy a survey.
      *
      * @access public
      * @return void
      * @throws CException
+     * @throws Exception
      */
     public function actionCopy()
     {
@@ -2220,131 +2223,61 @@ class SurveyAdministrationController extends LSBaseController
             App()->user->setFlash('error', gT("Access denied"));
             $this->redirect(App()->request->urlReferrer);
         }
-
-        $iSurveyID = sanitize_int(App()->request->getPost('copysurveylist'));
-
-        if (!Permission::model()->hasSurveyPermission($iSurveyID, 'surveycontent', 'export')) {
+        $surveyId = sanitize_int(App()->request->getPost('surveyIdToCopy'));
+        $survey = Survey::model()->findByPk($surveyId);
+        if (!$survey) {
+            App()->user->setFlash('error', gT("Survey does not exist."));
+            $this->redirect(App()->request->urlReferrer);
+        }
+        if (!Permission::model()->hasSurveyPermission($surveyId, 'surveycontent', 'export')) {
             App()->user->setFlash('error', gT("Access denied"));
             $this->redirect(App()->request->urlReferrer);
         }
 
+        //the new survey id
+        $rawNewSid = App()->request->getPost('copysurveyid');
+        $newSurveyId = ($rawNewSid === '' || $rawNewSid === null)
+            ? null
+            : sanitize_int($rawNewSid, 1, 999999);
+
+        //the source survey id
+        $sourceSurveyId = sanitize_int(App()->request->getPost('surveyIdToCopy'));
+
+        $sourceSurvey = Survey::model()->findByPk($sourceSurveyId);
+
+        $options = $this->getPostParamsForCopySurvey(App()->request);
+        $copySurveyService = new \LimeSurvey\Models\Services\CopySurvey(
+            $sourceSurvey,
+            $options,
+            $newSurveyId,
+        );
+
+        $copyResults = $copySurveyService->copy();
+
+        $copiedSurvey = $copyResults->getCopiedSurvey();
+        if ($copiedSurvey !== null) {
+            $groupList = QuestionGroup::model()->findAllByAttributes(['sid' => $copiedSurvey->sid]);
+            $this->resetExpressionManager($copiedSurvey, $groupList);
+        }
+
         $aData = [];
+        $aData['copyResults'] = $copyResults;
 
-        $aData['sHeader'] = gT("Copy survey");
-        $aData['sSummaryHeader'] = gT("Survey copy summary");
-        $aData['textCompleted'] = gT("Copy of survey is completed.");
-
-        // Start treatment and messagebox
-        $aData['bFailed'] = false;
-        $aExcludes = array();
-        if (App()->request->getPost('copysurveyexcludequotas') == "1") {
-            $aExcludes['quotas'] = true;
-        }
-
-        if (App()->request->getPost('copysurveyexcludepermissions') == "1") {
-            $aExcludes['permissions'] = true;
-        }
-
-        if (App()->request->getPost('copysurveyexcludeanswers') == "1") {
-            $aExcludes['answers'] = true;
-        }
-
-        if (App()->request->getPost('copysurveyresetconditions') == "1") {
-            $aExcludes['conditions'] = true;
-        }
-
-        if (App()->request->getPost('copysurveyresetstartenddate') == "1") {
-            $aExcludes['dates'] = true;
-        }
-
-        if (App()->request->getPost('copysurveyresetresponsestartid') == "1") {
-            $aExcludes['reset_response_id'] = true;
-        }
-
-        if (!$iSurveyID) {
-            $aData['sErrorMessage'] = gT("No survey ID has been provided. Cannot copy survey");
-            $aData['bFailed'] = true;
-        } elseif (!Survey::model()->findByPk($iSurveyID)) {
-            $aData['sErrorMessage'] = gT("Invalid survey ID");
-            $aData['bFailed'] = true;
-        } else {
-            App()->loadHelper('export');
-            $copysurveydata = surveyGetXMLData($iSurveyID, $aExcludes);
-            if (empty(App()->request->getPost('copysurveyname'))) {
-                $sourceSurvey = Survey::model()->findByPk($iSurveyID);
-                $sNewSurveyName = $sourceSurvey->currentLanguageSettings->surveyls_title;
-            } else {
-                $sNewSurveyName = App()->request->getPost('copysurveyname');
-            }
-        }
-
-        App()->loadHelper('admin.import');
-        if (!$aData['bFailed']) {
-            $copyResources = App()->request->getPost('copysurveytranslinksfields') == '1';
-            $translateLinks = $copyResources;
-            $aImportResults = XMLImportSurvey(
-                '',
-                $copysurveydata,
-                $sNewSurveyName,
-                sanitize_int(App()->request->getParam('copysurveyid'), '1', '999999'),
-                $translateLinks
-            );
-            if (isset($aExcludes['conditions'])) {
-                Question::model()->updateAll(array('relevance' => '1'), 'sid=' . $aImportResults['newsid']);
-                QuestionGroup::model()->updateAll(array('grelevance' => '1'), 'sid=' . $aImportResults['newsid']);
-            }
-
-            if (isset($aExcludes['reset_response_id'])) {
-                $oSurvey = Survey::model()->findByPk($aImportResults['newsid']);
-                $oSurvey->autonumber_start = 0;
-                $oSurvey->save();
-            }
-
-            if (!isset($aExcludes['permissions'])) {
-                Permission::model()->copySurveyPermissions($iSurveyID, $aImportResults['newsid']);
-            }
-
-            if (!empty($aImportResults['newsid']) && $copyResources) {
-                $resourceCopier = new CopySurveyResources();
-                [, $errorFilesInfo] = $resourceCopier->copyResources($iSurveyID, $aImportResults['newsid']);
-                if (!empty($errorFilesInfo)) {
-                    $aImportResults['importwarnings'][] = gT("Some resources could not be copied from the source survey");
-                }
-            }
-        } else {
-            $aData['bFailed'] = true;
-        }
-
-        // If the import failed, set the status and error message in order to keep consistency with other errors
-        if (!empty($aImportResults['error'])) {
-            $aData['sErrorMessage'] = $aImportResults['error'];
-            $aData['bFailed'] = true;
-        }
-
-        if (!$aData['bFailed'] && isset($aImportResults)) {
-            $aData['aImportResults'] = $aImportResults;
-            if (isset($aImportResults['newsid'])) {
-                // Set link pointing to survey administration overview. This link will be updated if the survey has groups
-                $aData['sLink'] = $this->createUrl('surveyAdministration/view/', ['iSurveyID' => $aImportResults['newsid']]);
-                $aData['sLinkApplyThemeOptions'] = 'surveyAdministration/applythemeoptions/surveyid/' . $aImportResults['newsid'];
-            }
-        }
-        if (!empty($aImportResults['newsid'])) {
-            $oSurvey = Survey::model()->findByPk($aImportResults['newsid']);
-            $aGrouplist = QuestionGroup::model()->findAllByAttributes(['sid' => $aImportResults['newsid']]);
-
-            $this->resetExpressionManager($oSurvey, $aGrouplist);
+        if ($copiedSurvey !== null) {
+            $aData['sLink'] = $this->createUrl('surveyAdministration/view/', ['iSurveyID' => $copiedSurvey->sid]);
+            $aData['sLinkApplyThemeOptions'] = 'surveyAdministration/applythemeoptions/surveyid/' . $copiedSurvey->sid;
+            $questionGroupList = QuestionGroup::model()->findAllByAttributes(['sid' => $copiedSurvey->sid]);
 
             // Make the link point to the first group/question if available
-            if (!empty($aGrouplist)) {
-                $oFirstGroup = $aGrouplist[0];
+            if (!empty($questionGroupList)) {
+                $oFirstGroup = $questionGroupList[0];
                 $oFirstQuestion = Question::model()->primary()->findByAttributes(
                     ['gid' => $oFirstGroup->gid],
                     ['order' => 'question_order ASC']
                 );
 
                 $aData['sLink'] = $this->getSurveyAndSidemenueDirectionURL(
-                    $aImportResults['newsid'],
+                    $copiedSurvey->sid,
                     $oFirstGroup->gid,
                     !empty($oFirstQuestion) ? $oFirstQuestion->qid : null,
                     'structure'
@@ -2352,16 +2285,108 @@ class SurveyAdministrationController extends LSBaseController
             }
         }
 
-        if ((App()->getConfig("editorEnabled")) && isset($aImportResults['newsid'])) {
-            if (!isset($oSurvey)) {
-                $oSurvey = Survey::model()->findByPk($aImportResults['newsid']);
-            }
-            if ($oSurvey->getTemplateEffectiveName() == 'fruity_twentythree') {
-                $aData['sLink'] = App()->createUrl("editorLink/index", ["route" => "survey/" . $aImportResults['newsid']]);
+        if ((App()->getConfig("editorEnabled")) && ($copiedSurvey !== null)) {
+            $copiedSurvey->setOptions();
+            if ($copiedSurvey->getTemplateEffectiveName() == 'fruity_twentythree') {
+                $aData['sLink'] = App()->createUrl("editorLink/index", ["route" => "survey/" . $copiedSurvey->sid]);
             }
         }
         $this->aData = $aData;
-        $this->render('importSurvey_view', $this->aData);
+
+        $this->render('copySurveyResult_view', $this->aData);
+    }
+
+    /**
+     * Initialises the necessary options.
+     *
+     * @param $request
+     * @return CopySurveyOptions
+     */
+    private function getPostParamsForCopySurvey($request): CopySurveyOptions
+    {
+        $optionsDataContainer = new CopySurveyOptions();
+
+        //Survey resource files and adapt links
+        $option = $request->getPost('copyResourcesAndLinks');
+        $optionsDataContainer->setResourcesAndLinks(isset($option) && $option == "1");
+
+        $option = $request->getPost('copySurveyQuotas');
+        $optionsDataContainer->setQuotas(isset($option) && $option == "1");
+
+        $option = $request->getPost('copySurveyPermissions');
+        $optionsDataContainer->setPermissions(isset($option) && $option == "1");
+
+        $option = $request->getPost('copyAnswerOptions');
+        $optionsDataContainer->setAnswerOptions(isset($option) && $option == "1");
+
+        $option = $request->getPost('copySurveyConditions');
+        $optionsDataContainer->setConditions(isset($option) && $option == "1");
+
+        $option = $request->getPost('copyStartEndDate');
+        $optionsDataContainer->setStartAndEndDate(isset($option) && $option == "1");
+
+        $option = $request->getPost('resetResponseStartId');
+        $optionsDataContainer->setResetResponseStartId(isset($option) && $option == "1");
+
+        return $optionsDataContainer;
+    }
+
+    /**
+     * @param int $surveyIdToCopy
+     * @return void
+     * @throws Exception
+     */
+    public function actionCopySimple($surveyIdToCopy)
+    {
+        //everybody who has permission to create surveys
+        if (!Permission::model()->hasGlobalPermission('surveys', 'create')) {
+            App()->user->setFlash('error', gT("Access denied"));
+            $this->redirect(App()->request->urlReferrer);
+        }
+
+        $surveyId = sanitize_int($surveyIdToCopy);
+        $survey = Survey::model()->findByPk($surveyId);
+        if (!$survey) {
+            App()->user->setFlash('error', gT("Survey does not exist."));
+            $this->redirect(App()->request->urlReferrer);
+        }
+
+        if (!Permission::model()->hasSurveyPermission($surveyId, 'surveycontent', 'export')) {
+            App()->user->setFlash('error', gT("Access denied"));
+            $this->redirect(App()->request->urlReferrer);
+        }
+        $optionsDataContainer = new CopySurveyOptions();
+        //start and enddate should be copied in the simple copy process
+        $optionsDataContainer->setStartAndEndDate(true);
+        $copySurveyService = new \LimeSurvey\Models\Services\CopySurvey(
+            $survey,
+            $optionsDataContainer,
+        );
+
+        $copyResults = $copySurveyService->copy();
+
+        $copiedSurvey = $copyResults->getCopiedSurvey();
+        if ($copiedSurvey !== null) {
+            $groupList = QuestionGroup::model()->findAllByAttributes(['sid' => $copiedSurvey->sid]);
+            $this->resetExpressionManager($copiedSurvey, $groupList);
+        }
+
+        if (empty($copyResults->getErrors())) {
+            App()->user->setFlash('success', gT("Survey copied successfully."));
+        } else {
+            App()->user->setFlash('error', gT("Error while copying the survey."));
+        }
+
+        $redirectUrl = App()->request->urlReferrer;
+        if ($copiedSurvey !== null) {
+            $redirectUrl = App()->createUrl("surveyAdministration/view/", ["iSurveyID" => $copiedSurvey->sid]);
+            $copiedSurvey->setOptions();
+            if ((App()->getConfig("editorEnabled")) && $copiedSurvey->getTemplateEffectiveName() == 'fruity_twentythree') {
+                $redirectUrl = App()->createUrl("editorLink/index", ["route" => "survey/" . $copiedSurvey->sid]);
+            }
+        }
+
+        $this->redirect($redirectUrl);
     }
 
     public function actionImport()
