@@ -9,7 +9,7 @@
         name="question[<?= $this->generalOption->name; ?>]" 
         id="<?= CHtml::getIdByName($this->generalOption->name); ?>"
         <?= ($this->generalOption->formElement->help) ? 'aria-describedby="help-' . CHtml::getIdByName($this->generalOption->name) . '"' : "" ?>
-        ><?= $this->generalOption->formElement->value; ?></textarea>
+        ><?= CHtml::encode($this->generalOption->formElement->value); ?></textarea>
     <?php if (isset($this->generalOption->formElement->options['inputGroup']['suffix'])) : ?>
         <div class="input-group-text">
             <?= $this->generalOption->formElement->options['inputGroup']['suffix']; ?>
