@@ -1,6 +1,6 @@
 <?php
 // Translation-source for new editor.
-// Updated on 2025-11-17 07:53:43
+// Updated on 2025-11-28 11:37:00
 gT('%d answer options');
 gT('%s bytes');
 gT('1 answer option');
@@ -299,142 +299,40 @@ gT('Rotate');
 gT('Saving...');
 gT('Unable to save image settings. Please try again.');
 gT('Zoom');
-gT('Customize link');
-gT('%s Congrats! You have completed the tour.');
-gT('Add content');
-gT('All Data');
-gT('Allow survey embedding');
+gT('Adjust the color of buttons and the progress bar');
+gT('Adjust the color of buttons and the progress bar in dark mode');
+gT('Adjust the color of your survey questions');
+gT('Alternate cell background color for better detectability');
 gT('An update is available. You will be redirected to the admin panel to perform the update.');
-gT('Answer already exists');
 gT('Anyone with link');
 gT('Anyone with the link to this survey can access.');
-gT('Are you sure you want to delete the attachments for the selected responses?');
-gT('Are you sure you want to perform this action?');
-gT('Array five points');
-gT('Array increase, same, and decrease');
-gT('Array ten points');
-gT('Array yes, no, and uncertain');
-gT('Autosave');
-gT('Can not be used with active survey.');
-gT('Canvas');
-gT('Clear filters');
-gT('Click and type to fill in your question text. You can also use the toolbar for formatting and more options.');
-gT('Click on the survey title to edit it, your changes will be saved automatically.');
-gT('Close interactive help');
-gT('Complete responses');
-gT('Completed without answers');
-gT('Conditions written in ExpressionScript mode can’t be displayed inside the visual condition builder.');
-gT('Confirm Action');
-gT('Create surveys in seconds with LimeSurvey. Easy to use, secure, and trusted by professionals worldwide. Get started free and unlock fresh insights today!');
-gT('Currently embedding is disabled for the application');
-gT('Currently embedding is disabled for the survey');
-gT('Customize preview');
-gT('Customize the link preview for your survey');
-gT('Data view');
-gT('Date/Time Format');
-gT('Display code');
-gT('Do you have a minute?');
-gT('Download file');
-gT('Drop icon here');
-gT('Edit text');
-gT('Embed code');
-gT('Embed survey options');
-gT('Embed the survey easily on a website');
-gT('Enable embedding in the survey settings under publication and access.');
-gT('Export response');
-gT('Export statistics');
-gT('Export survey responses');
-gT('Failed to download QR code');
-gT('File has been deleted');
-gT('Filter column');
-gT('Filter responses');
-gT('Filtered Data');
+gT('Arrow right');
+gT('Asterisk');
+gT('Change or upload a new icon');
+gT('Change the background color of your survey');
+gT('Change the corner radius (rounding) of buttons');
+gT('Check');
+gT('Check circle');
+gT('Check square');
+gT('Chevron circle right');
+gT('Corner radius');
+gT('Cross-hover in array-questions');
+gT('Dark mode');
+gT('Display options');
 gT('Fix number values automatically');
-gT('Huge text');
-gT('Incompleted without answers');
-gT('Interactive help');
-gT('Invalid input');
-gT('Last action');
+gT('Full width');
+gT('Highlight the whole row and column on hover');
 gT('Link with access code');
-gT('List dropdown');
-gT('Loading preview');
-gT('Loading responses...');
-gT('Loading statistics...');
-gT('Max');
-gT('Max ID must be a positive integer');
-gT('Max ID must be greater than Min ID');
-gT('MB');
-gT('Meta title - short and precise');
-gT('Min ID must be a positive integer');
-gT('Min ID must be less than Max ID');
-gT('No files uploaded');
-gT('Not enabled');
-gT('Only completed');
+gT('Note:');
 gT('Only participants with the link and access code can access.');
-gT('Open the settings to access theme options and change your survey\'s design. Try it and change the style of your survey.');
-gT('Organize columns');
-gT('Overwriting condition from condition builder');
-gT('Overwriting condition from ExpressionScript');
-gT('Pin column');
-gT('Please enable the embedding in the Global settings under Security.');
-gT('Powered by LimeSurvey – The Freshest Online Survey Tool');
-gT('Preview Settings');
-gT('Random A-Z/Z-A');
-gT('Reorder questions');
-gT('Responses can only be updated through the responses list.');
-gT('Responses list');
-gT('Responses rate');
-gT('Responses updated!');
-gT('Results overview');
-gT('Return to the editor');
-gT('Rows per page');
-gT('See how LimeSurvey autosaves your work? You don\'t have to worry about saving your progress at any time.');
-gT('Seed');
-gT('Select a question type');
-gT('Select an embed type');
+gT('participants.');
+gT('Plus');
+gT('Resistance');
 gT('Select how you want to share your survey');
-gT('Select the plus symbol to add questions, groups or logic to the survey. For now, let\'s add a simple multiple-choice question.');
-gT('Select the question type for your new question from this menu. Try it and add a simple multiple-choice question for now.');
-gT('Select which columns you want to display inside this table. You can order the columns by drag and drop.');
-gT('selected');
-gT('Sharing options');
-gT('Sharing overview');
-gT('Show in search results');
-gT('Submit date');
-gT('Subquestions order');
-gT('Switching to the visual condition builder will overwrite the condition written in expression script mode.');
-gT('Tell us');
-gT('The condition created inside the builder will replace your existing condition written in ExpressionScript mode. This action cannot be undone. Do you want to apply your changes?');
-gT('The preview doesn’t update automatically while you’re editing. It will refresh once the survey is activated. Click “Update preview” to see the latest changes now.');
-gT('The public registration feature is not supported for embedded surveys. Participants will not see the registration screen; instead, they will see the token form when applicable. To use public registration, share the survey link instead.');
-gT('The theme options allow you to change the appearance of your survey. Try it and change a style of your survey.');
-gT('This is where you create your questions and answer options. A sample question group and question is always created to get you started. Just click on any element to edit it.');
-gT('This script will be overwritten:');
-gT('Thumbnail');
-gT('To');
-gT('Token');
-gT('Unpin column');
-gT('Update global settings');
-gT('Update preview');
-gT('Update survey settings');
-gT('Use the left-side menu bar to navigate back to your survey at any time.');
-gT('Use the survey preview to see how your survey will look like for your participants.');
-gT('Use the trash-icon to delete questions  or the X-icons to delete answer options.');
-gT('Use this panel to change question types, logic, and display settings. You can also switch between simple and advanced question settings.');
-gT('View all responses');
-gT('View results overview');
-gT('View/Edit response details');
-gT('We would love to hear your thoughts on our sharing options!');
-gT('Welcome to LimeSurvey');
-gT('When you embed your survey on an external site, the code snippet will be visible in the site’s source. Anyone can copy this code and place your survey outside of your intended site. We recommend using it only on sites you manage or fully trust. Your survey and responses will always remain securely stored in your account.');
-gT('When your survey is ready, use this button to activate it. Once activated, you can access the options for sharing your survey.');
-gT('XLS');
-gT('You can reorder your questions and groups by drag and drop inside the survey structure.');
-gT('You currently have \"View only\" access.');
-gT('You do not have participants yet. Create or restore archived participants under %sparticipants%s');
-gT('You need an email tool like gmail or zoho or an email marketing tool like Mailchimp or Hubspot to embed the survey in an email.');
-gT('You still have a participants list for your survey. You can remove it under %sparticipants%s');
-gT('You will see a preview of your changes on the right side.');
-gT('You\'re now using the LimeSurvey editor. This quick tutorial will guide you through what\'s most important in just a few steps.');
-gT('Your condition will no longer be editable in the visual builder. Are you sure you want to apply these changes?');
-gT('At least %s pixels');
+gT('Star');
+gT('Theme color');
+gT('This question uses a customized condition. If you create a condition using this editor the current customized condition will be overwritten.');
+gT('Times');
+gT('Use the full width or create a left/right border');
+gT('You do not have participants yet. Create or restore archived participants under ');
+gT('You still have a participants list for your survey. You can remove it under ');
