@@ -7,13 +7,13 @@ use Exception;
 use InvalidArgumentException;
 use LimeSurvey\Models\Services\SurveyStatistics\Charts\StatisticsChartInterface;
 use LimeSurvey\Models\Services\SurveyStatistics\StatisticsResponseFilters;
-use LimeSurvey\Models\Services\SurveyStatistics\Charts\Questions\Processors\{MultipleChoiceProcessor,
+use LimeSurvey\Models\Services\SurveyStatistics\Charts\Questions\Processors\{ArrayNumbersProcessor,
+    MultipleChoiceProcessor,
     SingleOptionMultipleChartsProcessor,
     TextProcessor,
     RankingProcessor,
     SingleOptionProcessor,
-    DualScaleProcessor
-};
+    DualScaleProcessor};
 use Question;
 use Yii;
 
@@ -32,6 +32,7 @@ class QuestionStatistics implements StatisticsChartInterface
         $this->factories = [
             Question::QT_M_MULTIPLE_CHOICE => fn() => new MultipleChoiceProcessor(),
             Question::QT_N_NUMERICAL => fn() => new MultipleChoiceProcessor(),
+            Question::QT_COLON_ARRAY_NUMBERS => fn() => new ArrayNumbersProcessor(),
             Question::QT_P_MULTIPLE_CHOICE_WITH_COMMENTS => fn() => new MultipleChoiceProcessor(),
             Question::QT_T_LONG_FREE_TEXT => fn() => new TextProcessor(),
             Question::QT_S_SHORT_FREE_TEXT => fn() => new TextProcessor(),
@@ -47,7 +48,6 @@ class QuestionStatistics implements StatisticsChartInterface
             Question::QT_Q_MULTIPLE_SHORT_TEXT => fn() => null,
             Question::QT_SEMICOLON_ARRAY_TEXT => fn() => null,
             Question::QT_X_TEXT_DISPLAY => fn() => null,
-            Question::QT_COLON_ARRAY_NUMBERS => fn() => null,
             Question::QT_O_LIST_WITH_COMMENT => fn() => null,
 
             // Single option with multiple graphs for each subquestion
