@@ -41,8 +41,8 @@
         <div class="mb-3">
             <label class="  form-label" for="emailsmtphost"><?php eT("SMTP host:"); ?></label>
             <div class="">
-                <input class="form-control" type='text' size='50' id='emailsmtphost' name='emailsmtphost' value="<?php echo htmlspecialchars((string) getGlobalSetting('emailsmtphost')); ?>"/>
-                <span class="hint"><?php printf(gT("Enter your hostname and port, e.g.: %s"), "smtp.example.org:25"); ?></span>
+                <input class="form-control" type='text' size='50' aria-describedby="email_lb" id='emailsmtphost' name='emailsmtphost' value="<?php echo htmlspecialchars((string) getGlobalSetting('emailsmtphost')); ?>"/>
+                <span  id="email_lb" class="hint"><?php printf(gT("Enter your hostname and port, e.g.: %s"), "smtp.example.org:25"); ?></span>
             </div>
         </div>
         <div class="mb-3">
@@ -82,7 +82,8 @@
                     'selectOptions' => [
                         "0" => gT("Off", 'unescaped'),
                         "1" => gT("On errors", 'unescaped'),
-                        "2" => gT("Always", 'unescaped')
+                        "2" => gT("Always", 'unescaped'),
+                        "3" => gT("Always with connection details", 'unescaped')
                     ]
                 ]); ?>
             </div>
