@@ -15,6 +15,7 @@
                 <div>
                     <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                         'name'          => 'surveyPreview_require_Auth',
+                        'ariaLabel'=> gT('Survey preview only for administration users:'),
                         'checkedOption' => App()->getConfig('surveyPreview_require_Auth'),
                         'selectOptions' => [
                             '1' => gT('On'),
@@ -30,6 +31,7 @@
                 <div>
                     <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                         'name'          => 'filterxsshtml',
+                        'ariaLabel'=> gT('Filter HTML for XSS:'),
                         'checkedOption' => App()->getConfig('filterxsshtml'),
                         'selectOptions' => [
                             '1' => gT('On'),
@@ -52,6 +54,7 @@
                 <div>
                     <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                         'name'          => 'disablescriptwithxss',
+                        'ariaLabel'=> gT('Disable question script for XSS restricted user:'),
                         'checkedOption' => App()->getConfig('disablescriptwithxss'),
                         'selectOptions' => [
                             '1' => gT('On'),
@@ -75,6 +78,7 @@
                 <div class="">
                     <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                         'name'          => 'usercontrolSameGroupPolicy',
+                        'ariaLabel'=> gT('Group member can only see own group:'),
                         'id'            => 'usercontrolSameGroupPolicy',
                         'checkedOption' => App()->getConfig('usercontrolSameGroupPolicy'),
                         'selectOptions' => [
@@ -94,6 +98,7 @@
                 <div>
                     <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                         'name'          => 'x_frame_options',
+                        'ariaLabel'=> gT('IFrame embedding allowed:'),
                         'checkedOption' => Yii::app()->getConfig('x_frame_options'),
                         'selectOptions' => [
                             "allow"      => gT("Allow", 'unescaped'),
@@ -113,6 +118,7 @@
                 <div>
                     <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                         'name'          => 'force_ssl',
+                        'ariaLabel'=> gT('Force HTTPS:'),
                         'checkedOption' => App()->getConfig('force_ssl'),
                         'selectOptions' => [
                             "on"  => gT("On", 'unescaped'),
@@ -150,7 +156,7 @@
                     <div class='form-text'><?php eT("List of IP addresses to exclude from the maximum login attempts check. Separate each IP address with a comma or a new line."); ?></div>
                 </div>
 
-                <div class="mb-3">
+                <div class="mb-3">                      
                     <label class="form-label" for='maxLoginAttempt'>
                         <?php eT("Maximum number of attempts:"); ?>
                     </label>
