@@ -27,11 +27,12 @@ echo viewHelper::getViewTestTag('participantsBlacklistControl');
                     <div class="row ls-space margin top-10 bottom-10">
                         <div class="mb-3">
                             <label class='form-label col-md-8'>
-                                <?php eT('Blocklist all current surveys for participant once the global field is set:'); ?>
+                                <?php eT('Blocklist all current surveys for participant once the global field is set'); ?>
                             </label>
                             <div class='col-md-3'>
                                 <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                                     'name'          => 'blacklistallsurveys',
+                                    'ariaLabel'     => gT('Blocklist all current surveys for participant once the global field is set'),
                                     'checkedOption' => $blacklistallsurveys === 'Y' ? '1' : 0,
                                     'selectOptions' => [
                                         '1' => gT('Yes'),
@@ -44,11 +45,12 @@ echo viewHelper::getViewTestTag('participantsBlacklistControl');
                     <div class="row ls-space margin top-10 bottom-10">
                         <div class="mb-3">
                             <label class='form-label col-md-8'>
-                                <?php eT('Blocklist participant for any new added survey once the global field is set:'); ?>
+                                <?php eT('Blocklist all newly created surveys for participant once the global field is set'); ?>
                             </label>
                             <div class='col-md-3'>
                                 <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                                     'name'          => 'blacklistnewsurveys',
+                                    'ariaLabel'     => gT('Blocklist all newly created surveys for participant once the global field is set'),
                                     'checkedOption' => $blacklistnewsurveys === 'Y' ? '1' : 0,
                                     'selectOptions' => [
                                         '1' => gT('Yes'),
@@ -61,11 +63,12 @@ echo viewHelper::getViewTestTag('participantsBlacklistControl');
                     <div class="row ls-space margin top-10 bottom-10">
                         <div class="mb-3">
                             <label class='form-label col-md-8'>
-                                <?php eT('Prevent blocklisted participants from being added to a survey:'); ?>
+                                <?php eT('Prevent blocklisted participants from being added to a survey'); ?>
                             </label>
                             <div class='col-md-3'>
                                 <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                                     'name'          => 'blockaddingtosurveys',
+                                    'ariaLabel'     => gT('Prevent blocklisted participants from being added to a survey'),
                                     'checkedOption' => $blockaddingtosurveys === 'Y' ? '1' : 0,
                                     'selectOptions' => [
                                         '1' => gT('Yes'),
@@ -78,11 +81,12 @@ echo viewHelper::getViewTestTag('participantsBlacklistControl');
                     <div class="row ls-space margin top-10 bottom-10">
                         <div class="mb-3">
                             <label class='form-label col-md-8'>
-                                <?php eT('Hide blocklisted participants:'); ?>
+                                <?php eT('Hide blocklisted participants'); ?>
                             </label>
                             <div class='col-md-3'>
                                 <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                                     'name'          => 'hideblacklisted',
+                                    'ariaLabel'     => gT('Hide blocklisted participants'),
                                     'checkedOption' => $hideblacklisted === 'Y' ? '1' : 0,
                                     'selectOptions' => [
                                         '1' => gT('Yes'),
@@ -95,11 +99,12 @@ echo viewHelper::getViewTestTag('participantsBlacklistControl');
                     <div class="row ls-space margin top-10 bottom-10">
                         <div class="mb-3">
                             <label class='form-label col-md-8'>
-                                <?php eT( 'Delete globally blocklisted participant from the database:'); ?>
+                                <?php eT( 'Delete globally blocklisted participant from the database'); ?>
                             </label>
                             <div class='col-md-3'>
                                 <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                                     'name'          => 'deleteblacklisted',
+                                    'ariaLabel'     => gT('Delete globally blocklisted participant from the database'),
                                     'checkedOption' => $deleteblacklisted === 'Y' ? '1' : 0,
                                     'selectOptions' => [
                                         '1' => gT('Yes'),
@@ -112,11 +117,12 @@ echo viewHelper::getViewTestTag('participantsBlacklistControl');
                     <div class="row ls-space margin top-10 bottom-10">
                         <div class="mb-3">
                             <label class='form-label col-md-8'>
-                                <?php eT('Allow participant to remove himself/herself from blocklist:'); ?>
+                                <?php eT('Allow participant to remove himself/herself from blocklist'); ?>
                             </label>
                             <div class='col-md-3'>
                                 <?php $this->widget('ext.ButtonGroupWidget.ButtonGroupWidget', [
                                     'name'          => 'allowunblacklist',
+                                    'ariaLabel'     => gT('Allow participant to remove himself/herself from blocklist'),
                                     'checkedOption' => $allowunblacklist === 'Y' ? '1' : 0,
                                     'selectOptions' => [
                                         '1' => gT('Yes'),
