@@ -26,11 +26,11 @@ echo viewHelper::getViewTestTag('importParticipants');
                 <div class="row ls-space margin top-25 bottom-25">
                     <div class="col-lg-6">
                         <div class="mb-3">
-                            <label for="characterset" id="characterset" class='form-label '>
+                            <label for="characterset" class='form-label '>
                                 <?php eT("Character set of file:"); ?>
                             </label>
                             <div class="col-12">
-                                <select name="characterset"  class="form-select">
+                                <select id="characterset" name="characterset" class="form-select">
                                     <?php
                                     foreach (aEncodingsArray() as $key => $encoding):
                                         ?>
@@ -46,7 +46,7 @@ echo viewHelper::getViewTestTag('importParticipants');
                     </div>
                     <div class="col-lg-6">
                         <div class="mb-3">
-                            <label for="separatorused" id="separatorused" class='form-label '>
+                            <label for="separatorused" class='form-label '>
                                 <?php eT("Separator used:"); ?>
                             </label>
                             <div class="col-12">
@@ -58,7 +58,7 @@ echo viewHelper::getViewTestTag('importParticipants');
                                 );
                                 ?>
 
-                                <select name="separatorused"  class="form-select">
+                                <select name="separatorused" id="separatorused" class="form-select">
                                     <option value="auto" selected="selected"><?php eT("(Autodetect)"); ?></option>
                                     <?php
                                     foreach ($separatorused as $key => $separator):
@@ -76,7 +76,7 @@ echo viewHelper::getViewTestTag('importParticipants');
                 <div class="row  ls-space margin top-25 bottom-25">
                     <div class="col-lg-6">
                         <div class="mb-3">
-                            <label for="filter" id="filter" class='form-label '>
+                            <label id="filter" class='form-label '>
                                 <?php
                                 eT("Filter blank email addresses:");
                                 ?>
@@ -97,7 +97,7 @@ echo viewHelper::getViewTestTag('importParticipants');
 
                 <div class="col-12 ls-space margin top-25 bottom-25">
                     <div class="card card-primary">
-                        <div class="card-header ">
+                        <div role="heading" aria-level="2" class="card-header ">
                             <?php eT("CSV input format") ?>
                         </div>
                         <div class='card-body'>
