@@ -10,7 +10,7 @@
         <div class="col-12 content-center">
             <!-- Message box from super admin -->
             <div class="jumbotron message-box <?php echo $class ?? ""; ?>">
-                <div class="h2" role="heading" aria-level="2"><?php echo $title;?></div>
+                <h1 class="h2" ><?php echo $title;?></h1>
                 <?php echo $message;?>
             </div>
         </div>
