@@ -8,7 +8,7 @@
 ?>
 
 <div class='side-body'>
-    <h3><?php eT("Send email invitations"); ?></h3>
+    <h3 aria-level="1"><?php eT("Send email invitations"); ?></h3>
     <div class="row">
         <div class="col-12 content-right">
             <?php echo PrepareEditorScript(true, $this); ?>
