@@ -1,8 +1,11 @@
 <?php
-/**@var $modalTitle string  */
+/**@var $modalTitle string */
+
 ?>
 
 <div class="modal-header">
-    <h1 class="modal-title"><?php echo $modalTitle; ?></h1>
-    <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
+    <h5 class="modal-title">
+        <?php echo $modalTitle; ?>
+    </h5>
+    <button type="button" class="btn-close" data-bs-dismiss="modal" role="button"></button>
 </div>
