@@ -9,9 +9,9 @@ echo viewHelper::getViewTestTag('addQuota');
 <div class='side-body'>
     <div class="row">
         <div class="col-12 content-right">
-            <h3>
+            <h1 class="h3">
                 <?php eT("New quota");?>
-            </h3>
+            </h1>
             <?php $this->renderPartial('/quotas/_form',
                 array(
                     'oQuota'=>$oQuota,
