@@ -3,10 +3,9 @@
 namespace ls\tests\unit\api\opHandlers;
 
 use LimeSurvey\Libraries\Api\Command\V1\SurveyResponses\conditions\ContainConditionHandler;
-use PHPUnit\Framework\TestCase;
+use ls\tests\TestCondition;
 
-
-class OpHandlerResponseContainConditionTest extends TestCase
+class OpHandlerResponseContainConditionTest extends TestCondition
 {
     public function testCanHandleContain(): void
     {
@@ -32,8 +31,9 @@ class OpHandlerResponseContainConditionTest extends TestCase
         // Includes whitespace to verify value trimming inside execute()
         $criteria = $handler->execute('status', '  active  ');
 
+
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
-        $this->assertStringContainsString('`status` LIKE :match', $criteria->condition);
+        $this->assertFieldConditions($criteria->condition, '[0] LIKE :match', ['status']);
         $this->assertSame([':match' => '%active%'], $criteria->params);
     }
 
@@ -46,9 +46,11 @@ class OpHandlerResponseContainConditionTest extends TestCase
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
 
         // Expect OR chain with uniquely indexed placeholders
-        $this->assertStringContainsString('`first_name` LIKE :match0', $criteria->condition);
-        $this->assertStringContainsString(' OR ', $criteria->condition);
-        $this->assertStringContainsString('`last_name` LIKE :match1', $criteria->condition);
+        $this->assertFieldConditions(
+            $criteria->condition,
+            '[0] LIKE :match0 OR [1] LIKE :match1',
+            ['first_name', 'last_name']
+        );
 
         $this->assertSame(
             [
@@ -72,7 +74,11 @@ class OpHandlerResponseContainConditionTest extends TestCase
         $this->assertStringNotContainsString(';', $criteria->condition, 'Semicolon should be removed from condition.');
 
         // Expect the sanitized, quoted column name and LIKE placeholder
-        $this->assertStringContainsString('`nameDROPTABLEresponses--` LIKE :match', $criteria->condition);
+        $this->assertFieldConditions(
+            $criteria->condition,
+            '[0] LIKE :match',
+            ['nameDROPTABLEresponses--']
+        );
         $this->assertSame([':match' => '%ok%'], $criteria->params);
     }
 
@@ -85,8 +91,7 @@ class OpHandlerResponseContainConditionTest extends TestCase
         $handler = new ContainConditionHandler();
 
         $criteria = $handler->execute(['fieldA', 'fieldB', 'fieldC'], 'shared');
-
-        $pattern = '/`fieldA` LIKE :match0.*OR.*`fieldB` LIKE :match1.*OR.*`fieldC` LIKE :match2/s';
+        $pattern = '/(?:"|`|\[)?fieldA(?:"|`|\])?\s+LIKE\s+:match0.*?\s+OR\s+(?:"|`|\[)?fieldB(?:"|`|\])?\s+LIKE\s+:match1.*?\s+OR\s+(?:"|`|\[)?fieldC(?:"|`|\])?\s+LIKE\s+:match2/s';
         $this->assertTrue(
             (bool)preg_match($pattern, $criteria->condition),
             "Failed asserting OR chain uses unique indexed placeholders: {$criteria->condition}"
