@@ -3,10 +3,10 @@
 namespace ls\tests\unit\api\opHandlers;
 
 use LimeSurvey\Libraries\Api\Command\V1\SurveyResponses\conditions\DateRangeConditionHandler;
-use PHPUnit\Framework\TestCase;
+use ls\tests\TestCondition;
 
 
-class OpHandlerResponseDateRangeConditionTest extends TestCase
+class OpHandlerResponseDateRangeConditionTest extends TestCondition
 {
     public function testCanHandleDateRange(): void
     {
@@ -31,7 +31,11 @@ class OpHandlerResponseDateRangeConditionTest extends TestCase
         $criteria = $handler->execute('created_at', ['2024-07-01', '2024-07-31']);
 
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
-        $this->assertSame('`created_at` BETWEEN :created_atMin AND :created_atMax', $criteria->condition);
+        $this->assertFieldConditions(
+            $criteria->condition,
+            '[0] BETWEEN :created_atMin AND :created_atMax',
+            ['created_at']
+        );
         $this->assertSame(
             [
                 ':created_atMin' => '2024-07-01 00:00:00',
@@ -48,7 +52,7 @@ class OpHandlerResponseDateRangeConditionTest extends TestCase
         $criteria = $handler->execute('updated_at', ['2023-01-15', '']);
 
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
-        $this->assertSame('`updated_at` >= :updated_atMin', $criteria->condition);
+        $this->assertFieldConditions($criteria->condition, '[0] >= :updated_atMin', ['updated_at']);
         $this->assertSame([':updated_atMin' => '2023-01-15 00:00:00'], $criteria->params);
     }
 
@@ -59,7 +63,7 @@ class OpHandlerResponseDateRangeConditionTest extends TestCase
         $criteria = $handler->execute('updated_at', ['', '2023-01-31']);
 
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
-        $this->assertSame('`updated_at` <= :updated_atMax', $criteria->condition);
+        $this->assertFieldConditions($criteria->condition, '[0] <= :updated_atMax', ['updated_at']);
         $this->assertSame([':updated_atMax' => '2023-01-31 23:59:59'], $criteria->params);
     }
 
@@ -73,8 +77,7 @@ class OpHandlerResponseDateRangeConditionTest extends TestCase
 
         // Bad min date; good max date
         $criteria = $handler->execute('ts', ['07/01/2024', '2024-07-10']);
-
-        $this->assertSame('`ts` <= :tsMax', $criteria->condition);
+        $this->assertFieldConditions($criteria->condition, '[0] <= :tsMax', ['ts']);
         $this->assertSame([':tsMax' => '2024-07-10 23:59:59'], $criteria->params);
     }
 
@@ -84,8 +87,7 @@ class OpHandlerResponseDateRangeConditionTest extends TestCase
 
         // Good min date; bad max date
         $criteria = $handler->execute('ts', ['2024-07-01', '07-31-2024']);
-
-        $this->assertSame('`ts` >= :tsMin', $criteria->condition);
+        $this->assertFieldConditions($criteria->condition, '[0] >= :tsMin', ['ts']);
         $this->assertSame([':tsMin' => '2024-07-01 00:00:00'], $criteria->params);
     }
 
@@ -124,8 +126,11 @@ class OpHandlerResponseDateRangeConditionTest extends TestCase
         $criteria = $handler->execute('created_at`; DROP  table--', ['2024-06-01', '2024-06-02']);
 
         $this->assertStringNotContainsString(';', $criteria->condition);
-        $this->assertSame('`created_atDROPtable--` BETWEEN :created_atDROPtableMin AND :created_atDROPtableMax', $criteria->condition);
-
+        $this->assertFieldConditions(
+            $criteria->condition,
+            '[0] BETWEEN :created_atDROPtableMin AND :created_atDROPtableMax',
+            ['created_atDROPtable--']
+        );
         $this->assertSame(
             [
                 ':created_atDROPtableMin' => '2024-06-01 00:00:00',
