@@ -3,9 +3,9 @@
 namespace ls\tests\unit\api\opHandlers;
 
 use LimeSurvey\Libraries\Api\Command\V1\SurveyResponses\conditions\RangeConditionHandler;
-use PHPUnit\Framework\TestCase;
+use ls\tests\TestCondition;
 
-class OpHandlerResponseRangeConditionTest extends TestCase
+class OpHandlerResponseRangeConditionTest extends TestCondition
 {
     public function testCanHandleRange(): void
     {
@@ -32,8 +32,7 @@ class OpHandlerResponseRangeConditionTest extends TestCase
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
 
         // Condition should cast and use both Min and Max placeholders
-        $this->assertSame('CAST(`id` AS UNSIGNED) BETWEEN :idMin AND :idMax', $criteria->condition);
-
+        $this->assertFieldConditions($criteria->condition, 'CAST([0] AS UNSIGNED) BETWEEN :idMin AND :idMax', ['id']);
         $this->assertSame(
             [':idMin' => 10.0, ':idMax' => 25.0],
             $criteria->params
@@ -47,7 +46,7 @@ class OpHandlerResponseRangeConditionTest extends TestCase
         $criteria = $handler->execute('score', ['7', '']);
 
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
-        $this->assertSame('CAST(`score` AS UNSIGNED) >= :scoreMin', $criteria->condition);
+        $this->assertFieldConditions($criteria->condition, 'CAST([0] AS UNSIGNED) >= :scoreMin', ['score']);
         $this->assertSame([':scoreMin' => 7.0], $criteria->params);
     }
 
@@ -58,7 +57,7 @@ class OpHandlerResponseRangeConditionTest extends TestCase
         $criteria = $handler->execute('score', ['', '42']);
 
         $this->assertInstanceOf(\CDbCriteria::class, $criteria);
-        $this->assertSame('CAST(`score` AS UNSIGNED) <= :scoreMax', $criteria->condition);
+        $this->assertFieldConditions($criteria->condition, 'CAST([0] AS UNSIGNED) <= :scoreMax', ['score']);
         $this->assertSame([':scoreMax' => 42.0], $criteria->params);
     }
 
@@ -76,8 +75,11 @@ class OpHandlerResponseRangeConditionTest extends TestCase
         $criteria = $handler->execute('id`; DROP TABLE  responses--', ['1', '2']);
 
         $this->assertStringNotContainsString(';', $criteria->condition);
-
-        $this->assertSame('CAST(`idDROPTABLEresponses--` AS UNSIGNED) BETWEEN :idDROPTABLEresponsesMin AND :idDROPTABLEresponsesMax', $criteria->condition);
+        $this->assertFieldConditions(
+            $criteria->condition,
+            'CAST([0] AS UNSIGNED) BETWEEN :idDROPTABLEresponsesMin AND :idDROPTABLEresponsesMax',
+            ['idDROPTABLEresponses--']
+        );
 
         $this->assertArrayHasKey(':idDROPTABLEresponsesMin', $criteria->params);
         $this->assertArrayHasKey(':idDROPTABLEresponsesMax', $criteria->params);
@@ -128,7 +130,10 @@ class OpHandlerResponseRangeConditionTest extends TestCase
         $handler = new RangeConditionHandler();
 
         $criteria = $handler->execute('numeric_field', ['5', '15']);
-
-        $this->assertStringContainsString('CAST(`numeric_field` AS UNSIGNED)', $criteria->condition);
+        $this->assertFieldConditions(
+            $criteria->condition,
+            'CAST([0] AS UNSIGNED) BETWEEN :numeric_fieldMin AND :numeric_fieldMax',
+            ['numeric_field']
+        );
     }
 }
