@@ -94,6 +94,12 @@ class RemoteControlListParticipantsTest extends TestBaseClass
             ]
         ];
 
+        // for MSSQL Server
+        if (in_array(App()->db->driverName, ['mssql','sqlsrv'])) {
+            $list[0]['validuntil'] = preg_replace('/\.000$/', '', $list[0]['validuntil']);
+            $list[0]['validfrom'] = preg_replace('/\.000$/', '', $list[0]['validfrom']);
+        }
+
         $this->assertEquals($expected, $list);
     }
 
@@ -232,9 +238,133 @@ class RemoteControlListParticipantsTest extends TestBaseClass
         );
 
         $expected = [
-            'status' => 'Illegal operator: !'
+            'status' => 'Illegal operator: ! for column validuntil'
         ];
 
+        $this->assertEquals($expected, $list);
+    }
+
+    /**
+     * Test invalid columns 'extractvalue(1,concat(0x3a,(DATABASE())))'.
+     *
+     * @return void
+     */
+    public function testConditionInvalidColumn()
+    {
+        \Yii::import('application.helpers.remotecontrol.remotecontrol_handle', true);
+        \Yii::import('application.helpers.viewHelper', true);
+        \Yii::import('application.libraries.BigData', true);
+
+        // Create handler.
+        $admin   = new \AdminController('dummyid');
+        $handler = new \remotecontrol_handle($admin);
+
+        // Get session key.
+        $sessionKey = $handler->get_session_key(
+            self::$username,
+            self::$password
+        );
+        $this->assertNotEquals(['status' => 'Invalid user name or password'], $sessionKey);
+
+        /** @var array */
+        $list = $handler->list_participants(
+            $sessionKey,
+            self::$surveyId,
+            0,
+            10,
+            false,
+            false,
+            ['extractvalue(1,concat(0x3a,(DATABASE())))' => ['=', 1]]
+        );
+
+        $expected = [
+            'status' => 'Invalid column name: extractvalue(1,concat(0x3a,(DATABASE())))'
+        ];
+
+        $this->assertEquals($expected, $list);
+    }
+
+    /**
+     * Test simple string.
+     *
+     * @return void
+     */
+    public function testSimpleString()
+    {
+        \Yii::import('application.helpers.remotecontrol.remotecontrol_handle', true);
+        \Yii::import('application.helpers.viewHelper', true);
+        \Yii::import('application.libraries.BigData', true);
+
+        // Create handler.
+        $admin   = new \AdminController('dummyid');
+        $handler = new \remotecontrol_handle($admin);
+
+        // Get session key.
+        $sessionKey = $handler->get_session_key(
+            self::$username,
+            self::$password
+        );
+        $this->assertNotEquals(['status' => 'Invalid user name or password'], $sessionKey);
+
+        /** @var array */
+        $list = $handler->list_participants(
+            $sessionKey,
+            self::$surveyId,
+            0,
+            10,
+            false,
+            false,
+            ["validuntil >= 2020-04-01 15:12:00"]
+        );
+
+        /** @var array */
+        $expected = [
+            [
+                'tid' => "1",
+                'token' => "c",
+                'participant_info' => [
+                    'firstname' => "a",
+                    'lastname' => "b",
+                    'email' => "a@a.a"
+                ]
+            ]
+        ];
+        $this->assertEquals($expected, $list);
+    }
+
+    /**
+     * Test broken (SQL return DB Name) string.
+     * TODO : search other SQL payload
+     * @return void
+     */
+    public function testSQLPayloadString()
+    {
+        \Yii::import('application.helpers.remotecontrol.remotecontrol_handle', true);
+        \Yii::import('application.helpers.viewHelper', true);
+        \Yii::import('application.libraries.BigData', true);
+
+        // Create handler.
+        $admin   = new \AdminController('dummyid');
+        $handler = new \remotecontrol_handle($admin);
+
+        // Get session key.
+        $sessionKey = $handler->get_session_key(
+            self::$username,
+            self::$password
+        );
+        $this->assertNotEquals(['status' => 'Invalid user name or password'], $sessionKey);
+
+        /** @var array */
+        $list = $handler->mail_registered_participants(
+            $sessionKey,
+            self::$surveyId,
+            ['extractvalue(1,concat(0x3a,(DATABASE())))']
+        );
+
+        /** @var array */
+        $expected = [
+            'status' => 'Invalid expression for condition'
+        ];
         $this->assertEquals($expected, $list);
     }
 
@@ -309,6 +439,12 @@ class RemoteControlListParticipantsTest extends TestBaseClass
                 'validfrom' => "2020-03-18 15:12:00"
             ]
         ];
+
+        // for MSSQL Server
+        if (in_array(App()->db->driverName, ['mssql','sqlsrv'])) {
+            $list[0]['validuntil'] = preg_replace('/\.000$/', '', $list[0]['validuntil']);
+            $list[0]['validfrom'] = preg_replace('/\.000$/', '', $list[0]['validfrom']);
+        }
 
         $this->assertEquals($expected, $list);
 
